+++ app/registry/Component.js
@@ -1,4 +1,4 @@
-const joi = require('@hapi/joi');
+const joi = require('joi');
 const log = require('../log');
 
 /**