+++ app/triggers/providers/http/Http.test.js
@@ -1,4 +1,4 @@
-const { ValidationError } = require('@hapi/joi');
+const { ValidationError } = require('joi');
 const Http = require('./Http');
 
 const http = new Http();